Fix broken typedef generator. fix test.
